Rename rand_nodes to math_nodes
