+++ tests/tdastro/test_base_models.py
@@ -3,7 +3,8 @@ import random
 import jax
 import numpy as np
 import pytest
-from tdastro.base_models import FunctionNode, ParameterizedNode, ParameterSource, SingleVariableNode
+from tdastro.base_models import FunctionNode, ParameterizedNode, ParameterSource
+from tdastro.math_nodes.single_value_node import SingleVariableNode
 
 
 def _sampler_fun(**kwargs):
@@ -213,15 +214,6 @@ def test_parameterized_node_build_pytree():
     assert "A" not in pytree
     assert pytree["B"]["value1"] == 0.5
     assert pytree["B"]["value2"] == 3.0
-
-
-def test_single_variable_node():
-    """Test that we can create and query a SingleVariableNode."""
-    node = SingleVariableNode("A", 10.0)
-    assert str(node) == "SingleVariableNode"
-
-    state = node.sample_parameters()
-    assert node.get_param(state, "A") == 10
 
 
 def test_function_node_basic():
